refs #471 Get marker in navigator for all accounts
