+++ src/entities/marker.ts
@@ -1,0 +1,6 @@
+export type Marker = {
+  last_read_id: string
+  version: number
+  updated_at: string
+  unread_count?: number
+}
